connectorlist to own component
