+++ packages/connectkit/src/components/Pages/Connectors/styles.ts
@@ -1,13 +1,7 @@
 import styled from './../../../styles/styled';
-import { keyframes } from 'styled-components';
 
 import { motion } from 'framer-motion';
 import { ModalBody } from '../../Common/Modal/styles';
-
-const Shimmer = keyframes`
-  0%{ transform: translate(-100%) rotate(-45deg); }
-  100%{ transform: translate(100%) rotate(-80deg); }
-`;
 
 export const InfoBox = styled.div`
   padding: 24px 24px 28px;
@@ -80,238 +74,5 @@ export const LearnMoreButton = styled(motion.button)`
   }
   &:active {
     transform: scale(0.96);
-  }
-`;
-
-export const ConnectorsContainer = styled(motion.div)`
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-`;
-
-export const ConnectorButton = styled(motion.button)`
-  cursor: pointer;
-  user-select: none;
-  position: relative;
-  display: flex;
-  align-items: center;
-  padding: 0 20px;
-  width: 100%;
-  height: 64px;
-  font-size: 17px;
-  font-weight: var(--ck-primary-button-font-weight, 500);
-  line-height: 20px;
-  text-align: var(--ck-body-button-text-align, left);
-  transition: 180ms ease;
-  transition-property: background, color, box-shadow, transform;
-  will-change: transform, box-shadow, background-color, color;
-
-  --fallback-color: var(--ck-primary-button-color);
-  --fallback-background: var(--ck-primary-button-background);
-  --fallback-box-shadow: var(--ck-primary-button-box-shadow);
-  --fallback-border-radius: var(--ck-primary-button-border-radius);
-
-  --color: var(--ck-primary-button-color, var(--fallback-color));
-  --background: var(--ck-primary-button-background, var(--fallback-background));
-  --box-shadow: var(--ck-primary-button-box-shadow, var(--fallback-box-shadow));
-  --border-radius: var(
-    --ck-primary-button-border-radius,
-    var(--fallback-border-radius)
-  );
-
-  --hover-color: var(--ck-primary-button-hover-color, var(--color));
-  --hover-background: var(
-    --ck-primary-button-hover-background,
-    var(--background)
-  );
-  --hover-box-shadow: var(
-    --ck-primary-button-hover-box-shadow,
-    var(--box-shadow)
-  );
-  --hover-border-radius: var(
-    --ck-primary-button-hover-border-radius,
-    var(--border-radius)
-  );
-
-  --active-color: var(--ck-primary-button-active-color, var(--hover-color));
-  --active-background: var(
-    --ck-primary-button-active-background,
-    var(--hover-background)
-  );
-  --active-box-shadow: var(
-    --ck-primary-button-active-box-shadow,
-    var(--hover-box-shadow)
-  );
-  --active-border-radius: var(
-    --ck-primary-button-active-border-radius,
-    var(--hover-border-radius)
-  );
-
-  color: var(--color);
-  background: var(--background);
-  box-shadow: var(--box-shadow);
-  border-radius: var(--border-radius);
-
-  &:disabled {
-    transition: 180ms ease;
-  }
-
-  --bg: var(--background);
-  &:not(:disabled) {
-    &:hover {
-      color: var(--hover-color);
-      background: var(--hover-background);
-      box-shadow: var(--hover-box-shadow);
-      border-radius: var(--hover-border-radius);
-      --bg: var(--hover-background, var(--background));
-    }
-    &:focus-visible {
-      transition-duration: 100ms;
-      color: var(--hover-color);
-      background: var(--hover-background);
-      box-shadow: var(--hover-box-shadow);
-      border-radius: var(--hover-border-radius);
-      --bg: var(--hover-background, var(--background));
-    }
-    &:active {
-      color: var(--active-color);
-      background: var(--active-background);
-      box-shadow: var(--active-box-shadow);
-      border-radius: var(--active-border-radius);
-      --bg: var(--active-background, var(--background));
-    }
-  }
-`;
-export const ConnectorRecentlyUsed = styled(motion.span)`
-  position: relative;
-  top: var(--ck-recent-badge-top-offset, 0.5px);
-  display: inline-block;
-  padding: 10px 7px;
-  line-height: 0;
-  font-size: 13px;
-  font-weight: 400;
-  border-radius: var(--ck-recent-badge-border-radius, var(--border-radius));
-  color: var(
-    --ck-recent-badge-color,
-    var(--ck-accent-color, var(--ck-body-color-muted, currentColor))
-  );
-  background: var(--ck-recent-badge-background, transparent);
-  overflow: hidden;
-  span {
-    display: inline-block;
-    position: relative;
-  }
-  &:before {
-    z-index: 1;
-    content: '';
-    position: absolute;
-    inset: 0;
-    opacity: 0.4;
-    box-shadow: var(--ck-recent-badge-box-shadow, inset 0 0 0 1px currentColor);
-    border-radius: inherit;
-  }
-  &:after {
-    z-index: 2;
-    content: '';
-    position: absolute;
-    inset: -10%;
-    top: -110%;
-    aspect-ratio: 1/1;
-    opacity: 0.7;
-    background: linear-gradient(
-      170deg,
-      transparent 10%,
-      var(--ck-recent-badge-background, var(--bg)) 50%,
-      transparent 90%
-    );
-    animation: ${Shimmer} 2s linear infinite;
-  }
-`;
-
-export const ConnectorLabel = styled(motion.span)`
-  display: flex;
-  align-items: center;
-  gap: 9px;
-  width: 100%;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  padding: 2px 0;
-  padding-right: 38px;
-`;
-
-export const ConnectorIcon = styled(motion.div)`
-  position: absolute;
-  right: 20px;
-  width: 32px;
-  height: 32px;
-  overflow: hidden;
-  border-radius: 6px;
-  svg {
-    display: block;
-    width: 100%;
-    height: 100%;
-  }
-`;
-export const MobileConnectorsContainer = styled(motion.div)`
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  justify-content: space-between;
-  padding: 14px 0 28px;
-  margin: 0 0;
-`;
-
-export const MobileConnectorButton = styled(motion.button)`
-  --background: var(--ck-body-background-secondary);
-  cursor: pointer;
-  user-select: none;
-  position: relative;
-  padding: 0;
-  width: 100%;
-  min-width: 25%;
-  font-size: 13px;
-  font-weight: 500;
-  line-height: 13px;
-  text-align: center;
-  transition: transform 100ms ease;
-
-  background: none;
-
-  &:disabled {
-    cursor: not-allowed;
-    opacity: 0.4;
-  }
-
-  &:not(:disabled) {
-    &:active {
-      transform: scale(0.97);
-    }
-  }
-`;
-
-export const MobileConnectorLabel = styled(motion.span)`
-  display: block;
-  padding: 10px 0 0;
-  color: var(--ck-body-color);
-  opacity: 0.75;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-`;
-
-export const MobileConnectorIcon = styled(motion.div)`
-  margin: 0 auto;
-  width: 60px;
-  height: 60px;
-  overflow: hidden;
-  border-radius: 8px;
-  svg {
-    border-radius: inherit;
-    display: block;
-    position: relative;
-    transform: translate3d(0, 0, 0);
-    width: 100%;
-    height: 100%;
   }
 `;
